// backend/src/lib/api-zod/src/generated/api.ts
export const CreateContestantBody = zod.object({
  name: zod.string(),
  age: zod.number(),
  hometown: zod.string(),
  bio: zod.string(),
  zodiacSign: zod.string(),
  personality: zod.string(),
  interests: zod.string(),
  gender: zod.enum(["male", "female", "other"]),
});

/**
 * @summary Get a contestant by ID
 */
export const GetContestantParams = zod.object({
  id: zod.coerce.number(),
});
// ...
export const UpdateContestantParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateContestantBody = zod.object({
  name: zod.string(),
  age: zod.number(),
  hometown: zod.string(),
  bio: zod.string(),
  zodiacSign: zod.string(),
  personality: zod.string(),
  interests: zod.string(),
  gender: zod.enum(["male", "female", "other"]),
});
// ...
export const DeleteContestantParams = zod.object({
  id: zod.coerce.number(),
});